# swagger.json
{"openapi":"3.0.0","paths":{"/api/users/register":{"post":{"operationId":"UsersController_CreateUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"Registers a User","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/StandardResponse"},{"properties":{"data":{"type":"object","$ref":"#/components/schemas/CreateUserDto"}}},{"properties":{}}]}}}}},"tags":["Users"],"security":[{"bearer":[]}]}},"/api/users/login":{"post":{"operationId":"UsersController_LoginUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginUserDto"}}}},"responses":{"201":{"description":"Login a User","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/StandardResponse"},{"properties":{"data":{"type":"object","$ref":"#/components/schemas/LoginUserDto"}}},{"properties":{}}]}}}}},"tags":["Users"],"security":[{"bearer":[]}]}},"/api/wallet":{"get":{"operationId":"WalletController_getWallet","parameters":[],"responses":{"201":{"description":"View User Wallet Details","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/StandardResponse"},{"properties":{"data":{"type":"object","$ref":"#/components/schemas/Wallet"}}},{"properties":{}}]}}}}},"tags":["Wallet"],"security":[{"bearer":[]}]}},"/api/wallet/fund-me":{"post":{"operationId":"WalletController_fundMyAccount","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FundMyAccountDto"}}}},"responses":{"201":{"description":"Fund My Wallet","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/StandardResponse"},{"properties":{"data":{"type":"object","$ref":"#/components/schemas/Wallet"}}},{"properties":{}}]}}}}},"tags":["Wallet"],"security":[{"bearer":[]}]}},"/api/wallet/transfer-to-user":{"post":{"operationId":"WalletController_transferToUserByEmail","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransferAccountDto"}}}},"responses":{"201":{"description":"Transfer To User Wallet By Account Number","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/StandardResponse"},{"properties":{"data":{"type":"object","$ref":"#/components/schemas/Wallet"}}},{"properties":{}}]}}}}},"tags":["Wallet"],"security":[{"bearer":[]}]}},"/api/wallet/get-details/{accountNumber}":{"get":{"operationId":"WalletController_getDetails","parameters":[],"responses":{"201":{"description":"Fetch Wallet Details By Account Number","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/StandardResponse"},{"properties":{"data":{"type":"object","$ref":"#/components/schemas/Wallet"}}},{"properties":{}}]}}}}},"tags":["Wallet"],"security":[{"bearer":[]}]}},"/api/transactions/me":{"get":{"operationId":"TransactionsController_GetAllUserTransactions","parameters":[{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"take","required":false,"in":"query","schema":{"type":"number"}},{"name":"fromDate","required":false,"in":"query","schema":{"type":"string"}},{"name":"toDate","required":false,"in":"query","schema":{"type":"string"}},{"name":"search","required":false,"in":"query","schema":{"type":"string"}},{"name":"sortDirection","required":false,"in":"query","schema":{"type":"string"}},{"name":"transactionType","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"201":{"description":"Get a Users Transactions","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/StandardResponse"},{"properties":{"data":{"type":"object","$ref":"#/components/schemas/Transactions"}}},{"properties":{}}]}}}}},"tags":["Transactions"],"security":[{"bearer":[]}]}}},"info":{"title":"FintavaPay API","description":"FintavaPay API Documentation","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateUserDto":{"type":"object","properties":{}},"LoginUserDto":{"type":"object","properties":{}},"FundMyAccountDto":{"type":"object","properties":{}},"TransferAccountDto":{"type":"object","properties":{}},"PaginationMeta":{"type":"object","properties":{"page":{"type":"number"},"count":{"type":"number"},"skipped":{"type":"number"},"total":{"type":"number"}},"required":["page","count","skipped","total"]},"StandardResponse":{"type":"object","properties":{"success":{"type":"boolean"},"statusCode":{"type":"number"},"message":{"type":"string"},"data":{"type":"object"},"extra":{"type":"object"}},"required":["success","statusCode","message","data","extra"]},"TransactionQueryDto":{"type":"object","properties":{"page":{"type":"number"},"take":{"type":"number"},"fromDate":{"type":"string"},"toDate":{"type":"string"},"search":{"type":"string"},"sortDirection":{"type":"string"},"transactionType":{"type":"string"}}},"PaginationQueryDto":{"type":"object","properties":{"page":{"type":"number"},"take":{"type":"number"},"fromDate":{"type":"string"},"toDate":{"type":"string"},"search":{"type":"string"}}}}}}